+++ web/src/components/shared/CardContainer.tsx
@@ -18,7 +18,7 @@ export function CardContainer({
 
   const paddingClasses = {
     xs: "p-1 md:p-6",
-    sm: "p-3 md:p-5 lg:p-7",
+    sm: "p-4 lg:p-3",
     md: "p-4 sm:p-6",
     lg: "p-6 sm:p-8",
   }[padding];